+++ components/frontend/src/metric/MeasurementDetails.js
@@ -1,14 +1,33 @@
 import React from 'react';
-import { Tab, Menu } from 'semantic-ui-react';
+import { Dropdown, Icon, Tab, Menu } from 'semantic-ui-react';
 import { TrendGraph } from './TrendGraph';
 import { ReadOnlyOrEditable } from '../context/ReadOnly';
 import { Sources } from '../source/Sources';
 import { SourceEntities } from '../source/SourceEntities';
 import { MetricParameters } from './MetricParameters';
 import { FocusableTab } from '../widgets/FocusableTab';
-import { CopyButton, DeleteButton, MoveButtonGroup } from '../widgets/Button';
-import { copy_metric, delete_metric, set_metric_attribute } from '../api/metric';
+import { ItemActionButtons } from '../widgets/Button';
+import { ItemBreadcrumb } from '../widgets/ItemBreadcrumb';
+import { copy_metric, delete_metric, move_metric, set_metric_attribute } from '../api/metric';
 import { ChangeLog } from '../changelog/ChangeLog';
+import { get_source_name, get_subject_name } from '../utils';
+
+function subject_options(reports, datamodel, current_subject_uuid) {
+  let subject_options = [];
+    reports.forEach((report) => {
+      Object.entries(report.subjects).forEach(([subject_uuid, subject]) => {
+        const subject_name = get_subject_name(subject, datamodel);
+        subject_options.push({
+          content: <ItemBreadcrumb report={report.title} subject={subject_name} />,
+          disabled: subject_uuid === current_subject_uuid, key: subject_uuid,
+          text: report.title + subject_name,
+          value: subject_uuid
+        })
+      })
+    });
+    subject_options.sort((a, b) => a.text.localeCompare(b.text));
+    return subject_options;
+}
 
 export function MeasurementDetails(props) {
   const metric = props.report.subjects[props.subject_uuid].metrics[props.metric_uuid];
@@ -20,20 +39,10 @@ export function MeasurementDetails(props) {
       if (!report_source) { return }  // source was deleted, continue
       const nr_entities = (source.entities && source.entities.length) || 0;
       if (nr_entities === 0) { return } // no entities to show, continue
-      const source_type = report_source.type;
-      const source_name = report_source.name || props.datamodel.sources[source_type].name;
+      const source_name = get_source_name(report_source, props.datamodel);
       panes.push({
         menuItem: <Menu.Item key={source.source_uuid}><FocusableTab>{source_name}</FocusableTab></Menu.Item>,
-        render: () => <Tab.Pane>
-          <SourceEntities
-            datamodel={props.datamodel}
-            fetch_measurement_and_reload={props.fetch_measurement_and_reload}
-            metric={metric}
-            metric_uuid={props.metric_uuid}
-            report_uuid={report_uuid}
-            source={source}
-          />
-        </Tab.Pane>
+        render: () => <Tab.Pane><SourceEntities metric={metric} report_uuid={report_uuid} source={source} {...props} /></Tab.Pane>
       });
     });
   }
@@ -42,9 +51,7 @@ export function MeasurementDetails(props) {
     panes.push(
       {
         menuItem: <Menu.Item key='trend'><FocusableTab>{'Trend'}</FocusableTab></Menu.Item>,
-        render: () => <Tab.Pane>
-          <TrendGraph measurements={props.measurements} unit={unit_name} scale={props.scale} title={props.metric_name} />
-        </Tab.Pane>
+        render: () => <Tab.Pane><TrendGraph unit={unit_name} title={props.metric_name} {...props} /></Tab.Pane>
       }
     );
   }
@@ -52,13 +59,7 @@ export function MeasurementDetails(props) {
     {
       menuItem: <Menu.Item key='metric'><FocusableTab>{'Metric'}</FocusableTab></Menu.Item>,
       render: () => <Tab.Pane>
-        <MetricParameters
-          datamodel={props.datamodel}
-          fetch_measurement_and_reload={props.fetch_measurement_and_reload}
-          metric={metric}
-          metric_uuid={props.metric_uuid}
-          reload={props.reload}
-        />
+        <MetricParameters metric={metric} {...props} />
         <ChangeLog report_uuid={report_uuid} timestamp={props.report.timestamp} metric_uuid={props.metric_uuid} />
       </Tab.Pane>
     }
@@ -66,48 +67,37 @@ export function MeasurementDetails(props) {
   panes.push(
     {
       menuItem: <Menu.Item key='sources'><FocusableTab>{'Sources'}</FocusableTab></Menu.Item>,
-      render: () => <Tab.Pane>
-        <Sources
-          datamodel={props.datamodel}
-          measurement={props.measurement}
-          metric_type={metric.type}
-          metric_unit={props.unit}
-          metric_uuid={props.metric_uuid}
-          reload={props.reload}
-          report={props.report}
-          sources={metric.sources}
-          changed_fields={props.changed_fields}
-        />
-      </Tab.Pane>
+      render: () => <Tab.Pane><Sources metric_type={metric.type} sources={metric.sources} {...props} /></Tab.Pane>
     }
   );
+
+  function Buttons() {
+    return (
+      <ReadOnlyOrEditable editableComponent={
+        <div style={{ marginTop: "20px" }}>
+          <ItemActionButtons
+            item_type="metric"
+            first_item={props.first_metric}
+            last_item={props.last_metric}
+            onCopy={() => copy_metric(props.metric_uuid, props.reload)}
+            onDelete={() => delete_metric(props.metric_uuid, props.reload)}
+            onMove={(subject_uuid) => move_metric(props.metric_uuid, subject_uuid, props.reload)}
+            onReorder={(direction) => {
+              props.stop_sort();
+              set_metric_attribute(props.metric_uuid, "position", direction, props.reload)
+            }}
+            options={subject_options(props.reports, props.datamodel, props.subject_uuid)}
+            reorder_header={<Dropdown.Header>Report <Icon name='right chevron'/>Subject</Dropdown.Header>}
+            slot="row"
+          />
+        </div>}
+      />
+    )
+  }
   return (
     <>
       <Tab panes={panes} />
-      <ReadOnlyOrEditable editableComponent={
-        <>
-          <CopyButton
-            item_type="metric"
-            onClick={() => copy_metric(props.metric_uuid, props.reload)}
-          />
-          <MoveButtonGroup
-            first={props.first_metric}
-            last={props.last_metric}
-            marginTop="10px"
-            moveable="metric"
-            onClick={(direction) => {
-              props.stop_sort();
-              set_metric_attribute(props.metric_uuid, "position", direction, props.reload)}
-            }
-            slot="row"
-          />
-          <DeleteButton
-            item_type='metric'
-            onClick={() => delete_metric(props.metric_uuid, props.reload)}
-            style={{ marginTop: "10px" }}
-          />
-        </>
-      } />
+      <Buttons />
     </>
   )
 }