+++ components/frontend/src/report/Report.js
@@ -28,8 +28,18 @@ function ReportDashboard(props) {
         <CardDashboard
             cards={subject_cards.concat(tag_cards)}
             initial_layout={props.report.layout || []}
-            save_layout={function(layout) {set_report_attribute(props.report.report_uuid, "layout", layout, props.reload)}}
+            save_layout={function (layout) { set_report_attribute(props.report.report_uuid, "layout", layout, props.reload) }}
         />
+    )
+}
+
+function ReportErrorMessage({ report_date }) {
+    return (
+        <Message warning size='huge'>
+            <Message.Header>
+                {report_date ? `Sorry, this report didn't exist at ${report_date}` : "Sorry, this report doesn't exist"}
+            </Message.Header>
+        </Message>
     )
 }
 
@@ -44,14 +54,9 @@ export function Report(props) {
         // Make sure we only filter by tags that are actually used in this report
         setTags(prev_tags => prev_tags.filter(tag => Object.keys(props.report.summary_by_tag || {}).includes(tag)))
     }, [props.report]);
+
     if (!props.report) {
-        return (
-            <Message warning size='huge'>
-                <Message.Header>
-                    {props.report_date ? `Sorry, this report didn't exist at ${props.report_date}` : "Sorry, this report doesn't exist"}
-                </Message.Header>
-            </Message>
-        )
+        return <ReportErrorMessage report_date={props.report_date} />
     }
     return (
         <>
@@ -72,6 +77,7 @@ export function Report(props) {
                 nr_measurements={props.nr_measurements}
                 reload={props.reload}
                 report={props.report}
+                reports={props.reports}
                 report_date={props.report_date}
                 search_string={props.search_string}
                 tags={tags}