+++ components/frontend/src/source/Source.js
@@ -1,16 +1,39 @@
 import React from 'react';
-import { Grid, Header, Icon, Message } from 'semantic-ui-react';
+import { Dropdown, Grid, Header, Icon, Message } from 'semantic-ui-react';
 import { SourceType } from './SourceType';
 import { SourceParameters } from './SourceParameters';
 import { StringInput } from '../fields/StringInput';
 import { Logo } from '../logos/Logo';
 import { ChangeLog } from '../changelog/ChangeLog';
-import { CopyButton, DeleteButton, MoveButtonGroup } from '../widgets/Button';
-import { copy_source, delete_source, set_source_attribute } from '../api/source';
+import { ItemActionButtons } from '../widgets/Button';
+import { ItemBreadcrumb } from '../widgets/ItemBreadcrumb';
+import { copy_source, delete_source, move_source, set_source_attribute } from '../api/source';
 import { ReadOnlyOrEditable } from '../context/ReadOnly';
+import { get_metric_name, get_subject_name } from '../utils';
 
 function select_sources_parameter_keys(changed_fields, source_uuid) {
     return changed_fields ? changed_fields.filter((field) => field.source_uuid === source_uuid).map((field) => field.parameter_key) : []
+}
+
+function metric_options(reports, datamodel, current_metric_uuid) {
+    let metric_options = [];
+    reports.forEach((report) => {
+        Object.values(report.subjects).forEach((subject) => {
+            const subject_name = get_subject_name(subject, datamodel);
+            Object.entries(subject.metrics).forEach(([metric_uuid, metric]) => {
+                const metric_name = get_metric_name(metric, datamodel);
+                metric_options.push({
+                    content: <ItemBreadcrumb report={report.title} subject={subject_name} metric={metric_name} />,
+                    disabled: metric_uuid === current_metric_uuid,
+                    key: metric_uuid,
+                    text: report.title + subject_name + metric_name,
+                    value: metric_uuid
+                })
+            })
+        });
+    });
+    metric_options.sort((a, b) => a.text.localeCompare(b.text));
+    return metric_options;
 }
 
 export function Source(props) {
@@ -44,27 +67,23 @@ export function Source(props) {
         )
     }
 
-    function ButtonRow() {
+    function ButtonGridRow() {
         return (
             <ReadOnlyOrEditable editableComponent={
                 <Grid.Row>
                     <Grid.Column>
-                        <CopyButton
+                        <ItemActionButtons
                             item_type='source'
-                            onClick={() => copy_source(props.source_uuid, props.reload)}
-                        />
-                        <MoveButtonGroup
-                            first={props.first_source}
-                            last={props.last_source}
-                            moveable="source"
-                            onClick={(direction) => {
+                            first_item={props.first_source}
+                            last_item={props.last_source}
+                            onCopy={() => copy_source(props.source_uuid, props.reload)}
+                            onDelete={() => delete_source(props.source_uuid, props.reload)}
+                            onMove={(metric_uuid) => move_source(props.source_uuid, metric_uuid, props.reload)}
+                            onReorder={(direction) => {
                                 set_source_attribute(props.source_uuid, "position", direction, props.reload)
                             }}
-                            slot="position"
-                        />
-                        <DeleteButton
-                            item_type='source'
-                            onClick={() => delete_source(props.source_uuid, props.reload)}
+                            options={metric_options(props.reports, props.datamodel, props.metric_uuid)}
+                            reorder_header={<Dropdown.Header>Report <Icon name='right chevron'/>Subject <Icon name='right chevron'/>Metric</Dropdown.Header>}
                         />
                     </Grid.Column>
                 </Grid.Row>}
@@ -133,7 +152,7 @@ export function Source(props) {
                 {props.connection_error && <ErrorMessage title="Connection error" message={props.connection_error} />}
                 {props.parse_error && <ErrorMessage title="Parse error" message={props.parse_error} />}
                 <ChangeLogRow />
-                <ButtonRow />
+                <ButtonGridRow />
             </Grid>
         </>
     )