+++ components/frontend/src/utils.js
@@ -1,0 +1,11 @@
+export function get_metric_name(metric, datamodel) {
+    return metric.name || datamodel.metrics[metric.type].name;
+}
+
+export function get_source_name(source, datamodel) {
+    return source.name || datamodel.sources[source.type].name;
+}
+
+export function get_subject_name(subject, datamodel) {
+    return subject.name || datamodel.subjects[subject.type].name;
+}