+++ components/server/src/database/reports.py
@@ -6,7 +6,7 @@ from typing import cast, Any, Dict, List, Optional, Tuple
 import pymongo
 from pymongo.database import Database
 
-from server_utilities.functions import iso_timestamp
+from server_utilities.functions import iso_timestamp, unique
 from server_utilities.type import Change, Color, MetricId, ReportId, SourceId, Status, SubjectId
 from .datamodels import latest_datamodel
 from .measurements import last_measurements
@@ -103,7 +103,9 @@ def changelog(database: Database, nr_changes: int, **uuids):
     old_delta_filter = {f"delta.{key}": value for key, value in uuids.items() if value}
     delta_filter = {"$or": [old_delta_filter, {"delta.uuids": {"$in": list(uuids.values())}}]}
     changes.extend(database.reports.find(filter=delta_filter, sort=sort_order, limit=nr_changes, projection=projection))
-    return sorted(changes, reverse=True, key=lambda change: change["timestamp"])[:nr_changes]
+    changes = sorted(changes, reverse=True, key=lambda change: change["timestamp"])
+    # Weed out possible duplicate entries because the user moves items between reports, both reports get the same delta
+    return list(unique(changes, lambda change: cast(Dict[str, str], change["delta"])["description"]))[:nr_changes]
 
 
 def _get_report_uuid(reports, subject_uuid: SubjectId) -> Optional[ReportId]: